# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Authentication API",
    "description": "API for user authentication",
    "version": "1.0.0"
  },
  "paths": {
...
    "/applications/user/{userId}": {
      "get": {
        "tags": ["Applications"],
        "summary": "Fetch applications by user ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "userId",
          "in": "path",
          "required": true,
          "type": "integer",
          "description": "User ID"
        }],
        "responses": {
          "200": {
            "description": "List of applications for the user",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "500": {
            "description": "Error fetching applications for user"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Fetch all users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [{
          "in": "body",
          "name": "user",
          "description": "New auser data",
          "schema": {
            "$ref": "#/definitions/UserCreateRequest"
          }
        }],
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email format"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "options": {
        "tags": ["Users"],
        "summary": "List allowed methods",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Allowed methods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "allowed_methods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Fetch a user by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update a user by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to update",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "body",
            "required": true,
            "name": "updates",
            "description": "Update user data",
            "schema": {
              "$ref": "#/definitions/UserUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "head": {
        "tags": ["Users"],
        "summary": "Check if a user exists",
        "description": "Checks if a user exists by their ID.",
        "operationId": "checkUserExists",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "ID of the user to check"
          }
        ],
        "responses": {
          "200": {
            "description": "User exists"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Partially update a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Updates specific fields of user data based on the provided user ID and request body.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to update",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "body",
            "name": "user",
            "description": "Partially update user data",
            "schema": {
              "$ref": "#/definitions/UserUpdateRequest"
            }    
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{id}/jobs": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all jobs a user has applied for",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/UserJob"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{id}/apply": {
      "post": {
        "tags": ["Users"],
        "summary": "Apply for a job as a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to apply for the job",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "body",
            "description": "ID of the user applying for the job",
            "schema": {
              "$ref": "#/definitions/JobApplyRequest"
            }
          }
        ],
        "responses": {
           "201": {
            "description": "Application submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Application submitted successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Application": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "job_id": {
          "type": "integer"
        },
        "application_date": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "enum": ["pending", "accepted", "rejected"]
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "enum": ["admin", "user", "guest"]
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "UserCreateRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "enum": ["admin", "user", "guest"]
        }
      },
      "required": ["username", "email", "password"]
    },
    "UserUpdateRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "enum": ["admin", "user", "guest"]
        }
      }
    },
    "UserJob": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "JobApplyRequest": {
      "type": "object",
      "properties": {
        "job_id": {
          "type": "integer"
        }
      },
      "required": ["job_id"]
    }  
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format: Bearer token",
      "x-display-name": "Bearer Token",
      "x-token-prefix": "Bearer "
    }
  }
}